requirements.txt
streamlit
easyocr
pillow
pandas
numpy
